+++ okitweb/static/okit/js/okit_console.js
@@ -5,8 +5,8 @@
 console.info('Loaded Console Javascript');
 
 
-const okitVersion = '0.28.0';
-const okitReleaseDate = '20th October 2021';
+const okitVersion = '0.29.0';
+const okitReleaseDate = '10th November 2021';
 
 // Validation
 const validate_error_colour = "#ff4d4d";